Huge GUI Push!
So this project is going from doing some script-y analysis to a full GUI
that will provide data management, image processing tools, and metadata
storage.
This commit gives a primary GUI, project loading and data import
functionality. There's gonna be a whack load of changes, but this is at
least a primary stab.