Harry's change to the program up to date
Several changes made to the program. The Mueller matrix is first
converted to physical realizable by eliminating negative eigenvalue of
its covariance matrix before conducting decomposition. Parameters such
as Anisotropy (both linear and circular) and Asymmetric parameters are
added to the program. The orientation parameter of the linear
retardance(serial decomposition) and diattenuation are added.